FIXED crazy numbers problem
all seems correct now but not sure if its updating current velocity cause delta velocity is too small

+++ Source/HonsProject/Private/GravBody.cpp
@@ -89,7 +89,7 @@ void AGravBody::Tick(float DeltaTime)
 
 }
 
-void AGravBody::MoveBody(float editedDT)
+void AGravBody::MoveBody(double editedDT)
 {
 	SceneComponent->AddWorldOffset(velocity * editedDT);
 }
